requirements.txt
aiogram==3.*
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.5.3
openai==1.12.0